src/functions/requirements.txt
# Azure Functions Core
azure-functions==1.18.0

# Azure SDKs
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-mgmt-compute==30.4.0
azure-mgmt-network==25.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-securityinsight==2.0.0b2
azure-monitor-query==1.2.0
azure-storage-blob==12.19.0

# Microsoft Graph
msgraph-sdk==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# HTTP & API
requests==2.31.0
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Monitoring
opencensus-ext-azure==1.1.13

# Utilities
python-dateutil==2.8.2
cryptography==42.0.0
